scripts/people_connector/pyproject.toml
[project]
name = "people-data-import"
version = "0.1.0"
description = "A python project for people data import with Gemini Enterprise."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk==1.15.1",
    "python-dotenv>=1.0.0",
]

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"